Refactor broker config into ES module
