+++ server/src/trade/broker/broker.config.js
@@ -1,0 +1,9 @@
+export default {
+    commands: {
+        BUY: 'buy',
+        SELL: 'sell',
+        GET_POSITIONS: 'getPositions',
+        GET_POSITION: 'getPosition',
+        GET_ACCOUNT_INFO: 'getAccountInfo'
+    }
+};
