--- server/src/trade/broker/broker.config.json
@@ -1,9 +1,0 @@
-{
-    "commands": {
-        "BUY": "buy",
-        "SELL": "sell",
-        "GET_POSITIONS": "getPositions",
-        "GET_POSITION": "getPosition",
-        "GET_ACCOUNT_INFO": "getAccountInfo"
-    }
-}
